Domain Search Rewrite Cleanup: Remove legacy Start domains step
